Extract similarity computation out of LexRank model
This is the lowest impact split to divide these responsibilities into
separate classes and create a seam for testing.  There's still a fair bit
of refactoring left to do in each class, in order to end up with
sensible organization and APIs on both sides.
